separate out acceptance tests, fix block name, fix console.warn in jest test for filterbox
