+++ tests/acceptance/acceptancesuites/acceptancesuite.js
@@ -1,24 +1,20 @@
-import UniversalPage from './pageobjects/universalpage';
-import VerticalPage from './pageobjects/verticalpage';
-import { setupServer, shutdownServer } from './server';
-import FacetsPage from './pageobjects/facetspage';
+import UniversalPage from '../pageobjects/universalpage';
+import VerticalPage from '../pageobjects/verticalpage';
+import {
+  setupServer,
+  shutdownServer,
+  UNIVERSAL_PAGE,
+  FACETS_PAGE,
+  VERTICAL_PAGE
+} from '../server';
+import FacetsPage from '../pageobjects/facetspage';
 import { Selector, RequestLogger } from 'testcafe';
 import {
   browserBackButton,
   browserRefreshPage,
-  browserForwardButton,
   registerIE11NoCacheHook
-} from './utils';
-import {
-  expectRequestFiltersToEql,
-  expectRequestLocationRadiusToEql,
-  expectRequestDoesNotContainParam,
-  getMostRecentQueryParamsFromLogger
-} from './requestUtils';
-
-const UNIVERSAL_PAGE = 'http://localhost:9999/tests/acceptance/fixtures/html/universal';
-const VERTICAL_PAGE = 'http://localhost:9999/tests/acceptance/fixtures/html/vertical';
-const FACETS_PAGE = 'http://localhost:9999/tests/acceptance/fixtures/html/facets';
+} from '../utils';
+import { getMostRecentQueryParamsFromLogger } from '../requestUtils';
 
 /**
  * This file contains acceptance tests for a universal search page.
@@ -96,7 +92,7 @@ test('spell check flow', async t => {
   await t.expect(pageNum).eql('Page 1');
 
   // Check that clicking spell check sends a queryTrigger=suggest url param
-  // TODO(oshi) investigate making this an integration test
+  // TODO(SLAP-1062) investigate making this an integration test
   const queryParams = await getMostRecentQueryParamsFromLogger(spellCheckLogger);
   const queryTriggerParam = queryParams.get('queryTrigger');
   await t.expect(queryTriggerParam).eql('suggest');
@@ -171,213 +167,6 @@ test(`Facets load on the page, and can affect the search`, async t => {
   await filterBox.applyFilters();
   actualResultsCount = await verticalResultsComponent.getResultsCountTotal();
   await t.expect(actualResultsCount).eql(initialResultsCount);
-});
-
-test(`radius filter filterbox works with back/forward navigation and page refresh`, async t => {
-  const radiusFilterLogger = RequestLogger({
-    url: /v2\/accounts\/me\/answers\/vertical\/query/
-  });
-  await t.addRequestHooks(radiusFilterLogger);
-  await registerIE11NoCacheHook(t);
-  const searchComponent = FacetsPage.getSearchComponent();
-  await searchComponent.enterQuery('all');
-  await searchComponent.submitQuery();
-
-  const filterBox = FacetsPage.getStaticFilterBox();
-  const radiusFilter = await filterBox.getFilterOptions('DISTANCE');
-
-  // Choose the 25 miles radius filter
-  await radiusFilter.toggleOption('25 miles');
-  const filterTags = Selector('.yxt-ResultsHeader-removableFilterTag');
-  await t.expect(filterTags.count).eql(1);
-  await expectRequestLocationRadiusToEql(radiusFilterLogger, 40233.6);
-  radiusFilterLogger.clear();
-
-  // Hit the back button
-  await browserBackButton();
-  await t.expect(filterTags.count).eql(0);
-  await expectRequestDoesNotContainParam(radiusFilterLogger, 'locationRadius');
-  radiusFilterLogger.clear();
-
-  // Hit the forward button
-  await browserForwardButton();
-  await t.expect(filterTags.count).eql(1);
-  await expectRequestLocationRadiusToEql(radiusFilterLogger, 40233.6);
-  radiusFilterLogger.clear();
-
-  // Refresh the page
-  await browserRefreshPage();
-  await t.expect(filterTags.count).eql(1);
-  await expectRequestLocationRadiusToEql(radiusFilterLogger, 40233.6);
-});
-
-test(`static filter filterbox works with back/forward navigation and page refresh`, async t => {
-  const filterBoxLogger = RequestLogger({
-    url: /v2\/accounts\/me\/answers\/vertical\/query/
-  });
-  await t.addRequestHooks(filterBoxLogger);
-  await registerIE11NoCacheHook(t);
-  const martyFilter = {
-    c_puppyPreference: {
-      $eq: 'Marty'
-    }
-  };
-  const frodoFilter = {
-    c_puppyPreference: {
-      $eq: 'Frodo'
-    }
-  };
-  const filterBox = FacetsPage.getStaticFilterBox();
-  const staticFilter = await filterBox.getFilterOptions('STATIC FILTERS');
-  const filterTags = Selector('.yxt-ResultsHeader-removableFilterTag');
-  const searchComponent = FacetsPage.getSearchComponent();
-  await searchComponent.enterQuery('all');
-  await searchComponent.submitQuery();
-
-  // Click the 'Marty' checkbox
-  // 2 filter tags should be rendered, 1 from the backend facet and 1 from the static filter
-  await staticFilter.toggleOption('Marty');
-  await expectRequestFiltersToEql(filterBoxLogger, martyFilter);
-  await t.expect(filterTags.count).eql(2);
-
-  // Click the 'Frodo' checkbox
-  // 2 filter tags should be rendered, 1 from the backend facet and 1 from the static filter
-  await staticFilter.toggleOption('Frodo');
-  await t.expect(filterTags.count).eql(4);
-  await expectRequestFiltersToEql(filterBoxLogger, {
-    $or: [martyFilter, frodoFilter]
-  });
-
-  // Hit the back button, see the 'Frodo' filter disappear
-  await browserBackButton();
-  await expectRequestFiltersToEql(filterBoxLogger, martyFilter);
-  await t.expect(filterTags.count).eql(2);
-
-  // Hit the back button, see the 'Marty' filter disappear
-  await browserBackButton();
-  await expectRequestFiltersToEql(filterBoxLogger, {});
-  await t.expect(filterTags.count).eql(0);
-
-  // Hit the forward button, see the 'Marty' filter applied again
-  await browserForwardButton();
-  await expectRequestFiltersToEql(filterBoxLogger, martyFilter);
-  await t.expect(filterTags.count).eql(2);
-
-  // Hit the forward button, see the 'Frodo' filter applied again
-  await browserForwardButton();
-  await t.expect(filterTags.count).eql(4);
-  await expectRequestFiltersToEql(filterBoxLogger, {
-    $or: [martyFilter, frodoFilter]
-  });
-
-  // Refresh the page
-  await browserRefreshPage();
-  await t.expect(filterTags.count).eql(4);
-  await expectRequestFiltersToEql(filterBoxLogger, {
-    $or: [martyFilter, frodoFilter]
-  });
-});
-
-test(`filtersearch works with back/forward navigation and page refresh`, async t => {
-  const filterSearchLogger = RequestLogger({
-    url: /v2\/accounts\/me\/answers\/vertical\/query/
-  });
-  await t.addRequestHooks(filterSearchLogger);
-  await registerIE11NoCacheHook(t);
-  const expectOnlyFilterTagToEql = async expectedText => {
-    await t.expect(filterTags.count).eql(1);
-    const filterTagText = await filterTags.nth(0).find(
-      '.yxt-ResultsHeader-removableFilterValue').innerText;
-    return t.expect(filterTagText).eql(expectedText);
-  };
-  const filterTags = Selector('.yxt-ResultsHeader-removableFilterTag');
-  const filterSearch = FacetsPage.getFilterSearch();
-  const virginiaFilter = {
-    'builtin.location': {
-      $eq: 'P-region.7919684583758790'
-    }
-  };
-  const newYorkFilter = {
-    'builtin.location': {
-      $eq: 'P-place.15278078705964500'
-    }
-  };
-  await t.expect(filterSearchLogger.requests.length).eql(0);
-
-  // Choose the 'Virginia, United States' filter option
-  await filterSearch.selectFilter('Virginia, United States');
-  await expectOnlyFilterTagToEql('Virginia, United States');
-  await expectRequestFiltersToEql(filterSearchLogger, virginiaFilter);
-  filterSearchLogger.clear();
-
-  // Choose the 'New York City, New York, United States' filter option
-  await filterSearch.selectFilter('New York City, New York, United States');
-  await expectOnlyFilterTagToEql('New York City, New York, United States');
-  await expectRequestFiltersToEql(filterSearchLogger, newYorkFilter);
-  filterSearchLogger.clear();
-
-  // Hit the back button, expect to be back at the 'Virginia' filter state
-  await browserBackButton();
-  await expectOnlyFilterTagToEql('Virginia, United States');
-  await expectRequestFiltersToEql(filterSearchLogger, virginiaFilter);
-  filterSearchLogger.clear();
-
-  // Test that refreshing the page will use the 'Virginia' filter
-  await browserRefreshPage();
-  await expectOnlyFilterTagToEql('Virginia, United States');
-  await expectRequestFiltersToEql(filterSearchLogger, virginiaFilter);
-  filterSearchLogger.clear();
-
-  // Hit the back button, expect to be back at the initial state with 0 results
-  await browserBackButton();
-  await t.expect(filterTags.count).eql(0);
-  await t.expect(Selector('.yxt-StandardCard-title').count).eql(0);
-  await t.expect(filterSearchLogger.requests.length).eql(0);
-
-  // Hit the forward button, expect to see the 'Virginia' filter applied
-  await browserForwardButton();
-  await expectOnlyFilterTagToEql('Virginia, United States');
-  await expectRequestFiltersToEql(filterSearchLogger, virginiaFilter);
-  filterSearchLogger.clear();
-
-  // Hit the forward button, expect to see the 'New York' filter applied
-  await browserForwardButton();
-  await expectOnlyFilterTagToEql('New York City, New York, United States');
-  await expectRequestFiltersToEql(filterSearchLogger, newYorkFilter);
-  filterSearchLogger.clear();
-});
-
-test(`pagination works with page navigation after selecting a filtersearch filter`, async t => {
-  const expectFilterTagIsVirginia = async () => {
-    await t.expect(filterTags.count).eql(1);
-    const filterTagText = await filterTags.nth(0).find(
-      '.yxt-ResultsHeader-removableFilterValue').innerText;
-    return t.expect(filterTagText).eql('New York City, New York, United States');
-  };
-  const filterTags = Selector('.yxt-ResultsHeader-removableFilterTag');
-  const paginationComponent = FacetsPage.getPaginationComponent();
-  const filterSearch = FacetsPage.getFilterSearch();
-
-  await filterSearch.selectFilter('New York City, New York, United States');
-  await paginationComponent.clickNextButton();
-  let pageNum = await paginationComponent.getActivePageLabelAndNumber();
-  await t.expect(pageNum).eql('Page 2');
-  await expectFilterTagIsVirginia();
-
-  await browserBackButton();
-  pageNum = await paginationComponent.getActivePageLabelAndNumber();
-  await t.expect(pageNum).eql('Page 1');
-  await expectFilterTagIsVirginia();
-
-  await browserForwardButton();
-  pageNum = await paginationComponent.getActivePageLabelAndNumber();
-  await t.expect(pageNum).eql('Page 2');
-  await expectFilterTagIsVirginia();
-
-  await browserRefreshPage();
-  pageNum = await paginationComponent.getActivePageLabelAndNumber();
-  await t.expect(pageNum).eql('Page 2');
-  await expectFilterTagIsVirginia();
 });
 
 test(`selecting a sort option and refreshing maintains that sort selection`, async t => {
