+++ tests/acceptance/server.js
@@ -20,3 +20,7 @@ export async function setupServer (ctx) {
 export async function shutdownServer (ctx) {
   ctx.server.close();
 }
+
+export const UNIVERSAL_PAGE = 'http://localhost:9999/tests/acceptance/fixtures/html/universal';
+export const VERTICAL_PAGE = 'http://localhost:9999/tests/acceptance/fixtures/html/vertical';
+export const FACETS_PAGE = 'http://localhost:9999/tests/acceptance/fixtures/html/facets';
